refactor: extract dashboard component
